requirements.txt
Flask
icrawler
Pillow
supabase
python-dotenv
